Git Submodules.
This project needs to be built around submodules as much as
possible.

This is essentially just some scaffolding around all the actual
application code that lives in varius places in the OSM ecosystem,
and the last thing we want is for there to be a whole bunch of
ad-hoc patches to those libraries that only live here.

I'm not saying there will be *no* ad-hoc patches, I'm just saying
they'll be kept to a bare minimum.

And when they do exist, they'll be manually patched over code
pulled from the proper upstream repo.