nginx: Simple nginx configuration.
The nginx container is the main entry point to the application.

It's going to proxy incoming HTTP requests between cgimap and
the Rails application.

There's also a static maintenance mode that can be kicked off by
adding files to the shared htdocs volume.

We'll use this for importing progress status and an initial
loading screen where there would otherwise be a Bad Gateway error.