osmosis container with import utilities.
This uses a custom Go libary (planetstream) to chunk up PBF files
consisting of a single file Blob and the main file header into
standalone files for segmented import.

The segments are streamed in one-blob chunks over HTTP using
Range request headers and the start byte of the last Blob to start
processing is saved to a state file in a mounted volume so that
imports can begin from that point should the process terminate
before a full import is complete.

Additionally, we're building osmosis from scratch here because
there's a patch applied to the ApidbWriter class to add a SQL
clause called ON CONFLICT DO NOTHING to ensure that reimporting
the same block won't cause a fatal error when there is a unique
constraint colision during the INSERT query.

Perhaps someday that patch can be moved up to the upstream
osmosis repository, but for now the Git submodule is locked to
the proper commit so that no conflicts should arise during
the patch process.

The final piece of this system is a static file generator that
places a status HTML page within the shared htdocs volume.

This will pop nginx into maintenance mode (based on the nginx
config) to stop any API writes while the import is in process,
and display a dynamic progress status to the user.

It still takes a very long time to import a full Planetfile, but
at least the user experience is as friendly as possible and the
entire set of data doesn't have to be downloaded beforehand.