Postgres/postgis/libpgosm.
This container has a bit of cruft in it from a lot of experimental
import performance tweaking and debugging in an AWS cloud deploy
use case which isn't really applicable here, as those have been
stripped out of the codebase for public release.

One hack in this set of code is that it makes a raw file request
to GitHub at the commit hash of the Rails submodule in order to
grab the database schema file and load it directly here.
This is because it's not really desirable to share a build context
between this container and the Rails container.

The Rails container does a schema migration during its entrypoint
routine, which should eliminate the risk of the schema file
drifting from the Rails submodule.