Rails container.
Pretty simple wrapper around the openstreetmap-website submodule,
with the config files rsynced over the application source code
during the entrypoint process.

The provided configuration should be sufficient for development
deploys.

There's no volume mount for the Rails application code, meaning
a rebuild will have to be done to test each new code change
within the submodule code.  This should be fairly fast, however,
as the Gemfile is added to the container before any of the
application code; only changes to the package list will require
waiting for a bundle install.